requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
openai>=1.0
pypdf>=4.0
python-docx>=1.0
openpyxl>=3.0
xlrd>=2.0
